chore :: 메인페이지 로직 수정
