+++ src/App.tsx
@@ -6,8 +6,20 @@ import { Layout } from './layout';
 import { MainPage, VotePage, SubmitPage, DashboardPage, NotFoundPage } from './pages';
 import { queryClient } from './lib/react-query';
 import { globalStyles } from './styles/global';
+import {useEffect} from "react";
 
 const App = () => {
+    useEffect(() => {
+        const params = new URLSearchParams(window.location.search);
+        const token = params.get('token');
+        const email = params.get('email');
+
+        if (token && email) {
+            localStorage.setItem('access_token', token);
+            localStorage.setItem('email', email);
+            window.history.replaceState({}, '', '/');
+        }
+    }, []);
   return (
     <QueryClientProvider client={queryClient}>
       <Global styles={globalStyles} />
